Ajout du devoir facultatif et modification exercice 13 chap 4

+++ python/exercices/chapitre_4/exercice_13.py
@@ -1,5 +1,5 @@
 from random import randint
-def obtenirResultatExamen(note: int) -> str:
+def obtenirResultatExamen(note: float) -> str:
 	if note < 10:
 		resultat: str = "Éliminé !!"
 	elif note < 12:
@@ -13,11 +13,11 @@ def obtenirResultatExamen(note: int) -> str:
 
 	return resultat
 
-assert obtenirResultatExamen(9) == "Éliminé !!", "ça fonctionne pas pour note < 10"
-assert obtenirResultatExamen(11) == "Passable !", "ça fonctionne pas pour note < 12"
-assert obtenirResultatExamen(13) == "Assez Bien.", "ça fonctionne pas pour note < 14"
-assert obtenirResultatExamen(15) == "Bien !", "ça fonctionne pas pour note < 16"
-assert obtenirResultatExamen(20) == "Très Bien !!", "ça fonctionne pas pour note >= 15"
+assert obtenirResultatExamen(9.0) == "Éliminé !!", "ça fonctionne pas pour note < 10"
+assert obtenirResultatExamen(11.0) == "Passable !", "ça fonctionne pas pour note < 12"
+assert obtenirResultatExamen(13.0) == "Assez Bien.", "ça fonctionne pas pour note < 14"
+assert obtenirResultatExamen(15.0) == "Bien !", "ça fonctionne pas pour note < 16"
+assert obtenirResultatExamen(20.0) == "Très Bien !!", "ça fonctionne pas pour note >= 15"
 
 for i in range(20):
 	print(obtenirResultatExamen(randint(0,20)))
